requirements.txt
numpy==1.24.*
scipy==1.10.*
grpcio==1.57.*
psycopg2-binary==2.9.*
